Feat:new router shared,add users,add purchase list

+++ controllers/purchaseListController.js
@@ -80,29 +80,9 @@ exports.update = async (req, res, next) => {
 
         if (Object.keys(req.body).length == 0) return next(new Error('Nenhum dado informado.'))
 
-        const validFields = ['name', 'color', 'icon', 'concluded', 'deleted', 'total', 'shared_ids'];
+        const validFields = ['name', 'color', 'icon', 'deleted', 'total'];
         if (!Object.entries(req.body).some(([key, data]) => data !== undefined && data !== null && data !== '' && validFields.includes(key))) {
-            return next(new Error('Pelo menos um dos campos válidos deve estar presente no objeto: name, color, icon, concluded, deleted, total'))
-        }
-
-        if (req.body.shared_ids) {
-            const list = await prisma.purchaseList.findUnique({
-                where: {
-                    id,
-                },
-            })
-
-            if (list) {
-                const check = list.shared_ids.length > 0 ? list.shared_ids.some((item) => item !== req.body.shared_ids) : true
-                if (check) {
-                    list.shared_ids.push(req.body.shared_ids)
-                    req.body.shared_ids = list.shared_ids
-                } else {
-                    return next(new Error('Usuario ja adicionado.'))
-                }
-            } else {
-                return next(new Error('Id da lista invalida.'))
-            }
+            return next(new Error('Pelo menos um dos campos válidos deve estar presente no objeto: name, color, icon, deleted, total'))
         }
 
         const update = await prisma.purchaseList.update({
@@ -125,14 +105,66 @@ exports.update = async (req, res, next) => {
         // Outros erros
         throw new Error(error)
     }
+}
 
+exports.shared = async (req, res, next) => {
+    try {
+        const { id } = req.query
+        if (!id) return next(new Error('Informe um id'));
+
+        const { shared_ids } = req.body
+        if (!shared_ids) return next(new Error('Informe o id do usuario.'))
+
+        // Verificar se a lista é existente
+        const list = await prisma.purchaseList.findUnique({
+            where: {
+                id,
+            },
+        })
+
+        const { userId } = parserToken(req.cookies.token)
+        if (userId !== list.created_by_id) return next(new Error('Você não possui permissão para compartilhar esta lista.'))
+        
+        // Verificar se o id do usuario ja esta adicionado na lista
+        const check = list.shared_ids.length > 0 ? list.shared_ids.some((item) => item !== req.body.shared_ids) : true
+        if (check) {
+            list.shared_ids.push(req.body.shared_ids)
+            req.body.shared_ids = list.shared_ids
+        } else {
+            return next(new Error('Usuario ja adicionado.'))
+        }
+
+        const update = await prisma.purchaseList.update({
+            where: {
+                id,
+            },
+            data: {
+                shared_ids: req.body.shared_ids
+            }
+        })
+
+        res.status(200).json({
+            success: true,
+            update
+        })
+    } catch (error) {
+        // Verifica se o erro é devido a um ID inválido
+        if (error.message.includes('Malformed ObjectID')) {
+            return next(new Error('ID da lista inválido. Verifique se o ID está no formato correto.'))
+        }
+
+        // Outros erros
+        throw new Error(error)
+    }
 }
 
 exports.destroy = async (req, res, next) => {
     try {
         const { id } = req.query
-        // Check
         if (!id) return next(new Error('Informe o id do usuário'));
+
+        const { userId } = parserToken(req.cookies.token)
+        if (userId !== list.created_by_id) return next(new Error('Você não possui permissão para deletar a lista.'))
 
         await prisma.purchaseList.delete({
             where: {
